// src/main/resources/angular/src/app/api/model/ticket.ts
/**
 * Dum API
 * This is the DUM API
 *
 * OpenAPI spec version: 1.0.0
 * Contact: kfiertek@nexusgeographics.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { BonusApplied } from './bonusApplied';
import { ZoneType } from './zoneType';


export interface Ticket { 
    ticketId: number;
    plateNumber: string;
    city?: string;
    userId?: number;
    /**
     * Name or email or company, ....
     */
    userDesc?: string;
    zoneType?: ZoneType;
    startDate?: Date;
    /**
     * If not set, the ticket is still valid. A stopDate null indicates that the ticket state is \"process pending\"
     */
    stopDate?: Date;
    /**
     * Time spent in seconds at the parking. Notice that it's not the stop time minus start time, as there are non counting periods like midday or weekends.
     */
    parkingTime?: string;
    /**
     * Maximum parking time
     */
    maximumParkingTime?: number;
    segmentId?: number;
    scheduleDesc?: string;
    parkingName?: string;
    bonusApplied?: BonusApplied;
    ticketType?: number;
    segmentCode?:string;
    price?:number;
    unpaid:boolean;
}