// src/main/resources/angular/src/app/globalVariables/globalVariables.ts
export const AgremiatType = [{
  id: 1,
  name: 'Autònomo'
},
{
  id: 2,
  name: 'Empresa'
}];

export const EstatsUsuari = [
  {
    value: 1,
    nom: 'GESTIO_OPERATIVA.PENDING_ACTIVATION'
  },
  {
    value: 2,
    nom: 'GESTIO_OPERATIVA.ACTIVE'
  },
  {
    value: 3,
    nom: 'GESTIO_OPERATIVA.BLOCKED'
  },
  {
    value: 5,
    nom: 'GESTIO_OPERATIVA.DELETED_USER'
  }
];

export const ETIQUETAS_ECO: String[] = [
  '0', 'EC', 'C', 'B'
];

export enum FUNCIONALITIES {
  USER_QUERY = 'USER_QUERY',
  USER_MANAGEMENT = 'USER_MANAGEMENT',
  OPERATIONS_QUERY = 'OPERATIONS_QUERY',
  OPERATIONS_MANAGEMENT = 'OPERATIONS_MANAGEMENT',
  USERBO_QUERY = 'USERBO_QUERY',
  USERBO_MANAGEMENT = 'USERBO_MANAGEMENT',
  WATCHER_QUERY = 'WATCHER_QUERY',
  WATCHER_MANAGEMENT = 'WATCHER_MANAGEMENT',
  AREA_QUERY = 'AREA_QUERY',
  AREA_MANAGEMENT = 'AREA_MANAGEMENT',
  GUILD_QUERY = 'GUILD_QUERY',
  GUILD_MANAGEMENT = 'GUILD_MANAGEMENT',
  NOTIFICATIONS_MANAGEMENT = 'NOTIFICATIONS_MANAGEMENT',
  SYSTEM_ADMIN = 'SYSTEM_ADMIN',
  VEHICLES_QUERY = 'VEHICLES_QUERY',
  VEHICLES_MANAGEMENT = 'VEHICLES_MANAGEMENT'
};

export enum PROFILES {
  TOT_ACTIVAT = 'Tot activat',
  AREA = 'AREA',
  ADMIN = 'Administrador',
  AT_CLIENT_BASIC = 'Atenció Client Bàsic',
  AT_CLIENT_OPERATIU = 'Atenció Client Operatiu',
  AMB_ADMIN = 'AMB Administrador',
  AMB_FINANCE = 'AMB Finances',
  AMB_QUERY = 'AMB Consulta'
}


export enum APLICATION {
  ZONA_DUM,
  ZONA_BUS
}